refactor: organize ResultsViewController into extensions
