+++ ScanPlan/ResultsViewController/ResultsViewController+Export.swift
