+++ ScanPlan/ResultsViewController/ResultsViewController+TableView.swift
